Headless servers no longer create player characters
An unintended consequence of the headless server being treated just like a player is that they also created an avatar and a microphone stream. I added a check in GameManager.cs to not create a player avatar and voice stream if the headless server is creating themselves.

Small changes:
Changed spacing in GameManager.cs to be unform with other files. Pending style guidelines.
GameManager.cs uses the right namespace