Updated API
API now contains a config file where all secrets and the IML path go. The config file goes in the Config folder and its structure will be defined in the Wiki at a later date.
The API now supports Linux and will automatically handle file paths appropriately.

The database connection will be made automatically on the first request and will automatically reconnect if the connection is lost.