Create codegen.php
A proof of concept demo page that generates a 6 digit number code and stores it in the user's session. The idea is that, for production, this code will be put in a database that an API endpoint can then utilize for logging in. This is currently running on the IML server and can be tested.