Authcode functionality added
A new API endpoint, LoginVR, exists. To use it, send it a json string containing the email associated with the account and the login code. If the login code is still valid, and the emails match, the endpoint will respond with the userID which can be saved and then recalled again to access sensitive endpoints.