Some Button Function
Currently has a single floating button that sends a user to a random room. There are also commented out sections of code whose goals are to collect room names to then create buttons using those names to send users to specific rooms. The issue currently working on this is finding a means to have the number of needed buttons instead of just hardcoding a solution.